# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroGestion API",
    "description": "API para gestionar cultivos de principio a fin: siembras, ciclos, conteo de plantas, vivero, riesgos y eventos adversos. Todos los errores usan el mismo formato: decida siempre con el campo `error`, nunca con `message`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/glosario": {
      "get": {
        "tags": [
          "2. Cuenta y ayuda"
        ],
        "summary": "Términos del agro explicados",
        "operationId": "glosario_api_v1_glosario_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "description": "Término a buscar",
              "title": "Q"
            },
            "description": "Término a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TerminoSalida"
                  },
                  "title": "Response Glosario Api V1 Glosario Get"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/politica": {
      "get": {
        "tags": [
          "2. Cuenta y ayuda"
        ],
        "summary": "Política de tratamiento de datos",
        "operationId": "politica_api_v1_politica_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoliticaSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cuenta/consentimiento": {
      "get": {
        "tags": [
          "2. Cuenta y ayuda"
        ],
        "summary": "Última autorización aceptada",
        "operationId": "ver_consentimiento_api_v1_cuenta_consentimiento_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentimientoSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "2. Cuenta y ayuda"
        ],
        "summary": "Aceptar el tratamiento de datos",
        "operationId": "aceptar_api_v1_cuenta_consentimiento_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentimientoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentimientoSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentimientoEntrada": {
        "properties": {
          "version_politica": {
            "type": "string",
            "title": "Version Politica"
          },
          "acepta_tratamiento": {
            "type": "boolean",
            "title": "Acepta Tratamiento"
          },
          "acepta_transferencia_ia": {
            "type": "boolean",
            "title": "Acepta Transferencia Ia",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version_politica",
          "acepta_tratamiento"
        ],
        "title": "ConsentimientoEntrada"
      },
      "ConsentimientoSalida": {
        "properties": {
          "version_politica": {
            "type": "string",
            "title": "Version Politica"
          },
          "aceptado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Aceptado En"
          },
          "acepta_transferencia_ia": {
            "type": "boolean",
            "title": "Acepta Transferencia Ia"
          }
        },
        "type": "object",
        "required": [
          "version_politica",
          "aceptado_en",
          "acepta_transferencia_ia"
        ],
        "title": "ConsentimientoSalida"
      },
...
      "DetalleError": {
        "properties": {
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          }
        },
        "type": "object",
        "required": [
          "campo",
          "mensaje",
          "tipo"
        ],
        "title": "DetalleError"
      },
...
      "ErrorRespuesta": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetalleError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status_code"
        ],
        "title": "ErrorRespuesta",
        "description": "Formato único de error: decida siempre con `error`, nunca con `message`."
      },
...
      "PoliticaSalida": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "datos_que_usamos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Datos Que Usamos"
          },
          "para_que": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Para Que"
          },
          "sus_derechos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sus Derechos"
          }
        },
        "type": "object",
        "required": [
          "version",
          "estado",
          "datos_que_usamos",
          "para_que",
          "sus_derechos"
        ],
        "title": "PoliticaSalida"
      },
...
      "TerminoSalida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "termino": {
            "type": "string",
            "title": "Termino"
          },
          "explicacion": {
            "type": "string",
            "title": "Explicacion"
          },
          "categoria": {
            "type": "string",
            "title": "Categoria"
          }
        },
        "type": "object",
        "required": [
          "id",
          "termino",
          "explicacion",
          "categoria"
        ],
        "title": "TerminoSalida"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}